refactor(arenamaster): split modules
